+++ CHANGELOG.rst
@@ -32,8 +32,10 @@ TODO: integration test with pandapower (see `pandapower/contingency/contingency.
 [0.10.4] 2025-xx-yy
 ----------------------
 - [DEPRECATED] python 3.7 builds will no longer be available
+- [FIXED] a bug in the import of the grid from pypowsybl
 - [ADDED] compatibility with python 3.14
 - [IMPROVED] the way to initialize the transformers from pypowsybl
+- [IMPROVED] possibility to load grid with phase shifter in pypowsybl
 
 [0.10.3] 2025-04-28
 ----------------------